requirements.txt
Django>=5.0.0
python-dotenv>=1.0.1
google-generativeai>=0.4.0
